sync sanity

+++ src/fluxa/time.h
@@ -20,7 +20,27 @@
 #ifndef SPIRALCORE_TIME
 #define SPIRALCORE_TIME
 
-static const double ONE_OVER_UINT_MAX = 1.0/ULONG_MAX;
+static const double ONE_OVER_UINT_MAX = 1.0/UINT_MAX;
+
+// This is the deal with time
+// --------------------------
+// 
+// We have two types of time dealt with here, unix timestamps
+// and ntp timestamps - this class is an ntp timestamp but is derived
+// from gettimeofday which produces unix timestamps.
+//
+// Unix timestamps
+// ---------------
+// Unix timestamps can either be 32 or 64 bits, but both correspond 
+// to seconds and microseconds (millioth of a second) since 1970. Not
+// sure what the benefit of microseconds being 64 bit is, other than
+// symmetry?
+// 
+// NTP timestamps
+// --------------
+// These are 2 X 32 bits, seconds and fractions of a second (using the
+// full unsigned 32 bit int range) from 1900, we use this for OSC
+// compatibility
 
 namespace spiralcore
 {
@@ -29,7 +49,7 @@ class time
 {
  public:
   time();
-  time(unsigned long s, unsigned long f) : seconds(s),fraction(f) {}
+  time(unsigned int s, unsigned int f) : seconds(s),fraction(f) {}
   void set_to_now();
   void set_from_posix(timeval tv);
   void inc_by_sample(unsigned long samples, unsigned long samplerate);
@@ -41,12 +61,12 @@ class time
   time &operator+=(double s);
   void print() const;
   double get_fraction() const { return fraction*ONE_OVER_UINT_MAX; }
-  void set_fraction(double s) { fraction = (long)(s*(double)ULONG_MAX); }
+  void set_fraction(double s) { fraction = (unsigned int)(s*(double)UINT_MAX); }
   bool is_empty() { return (!seconds && !fraction); }
   double get_difference(const time& other);
 
-  unsigned long seconds;
-  unsigned long fraction;
+  unsigned int seconds;
+  unsigned int fraction;
 };
 
 }
